Missing envvar returns configured default.
If the envvar is unset, System/getenv returns a nil, which was being
assoc'd in place. So you couldn't actually have a default in config,
that you *optionally* override with an envvar. It'd just always be nil.

This changes that so that if the envvar is unset, it won't assoc to the
location; if a key did exist there it will now return.